Create profiles of observed variables using two-step cluster analysis

Create profiles of observed variables using two-step cluster analysis

create_profiles_cluster(df, ..., n_profiles, to_center = FALSE,
  to_scale = FALSE, distance_metric = "squared_euclidean",
  linkage = "complete")

Arguments

df
with two or more columns with continuous variables

...
unquoted variable names separated by commas

n_profiles
The specified number of profiles to be found for the clustering solution

to_center
Boolean (TRUE or FALSE) for whether to center the raw data with M = 0

to_scale
Boolean (TRUE or FALSE) for whether to scale the raw data with SD = 1

distance_metric
Distance metric to use for hierarchical clustering; "squared_euclidean" is default but more options are available (see ?hclust)

linkage
Linkage method to use for hierarchical clustering; "complete" is default but more options are available (see ?dist)

Value

A list containing the prepared data, the output from the hierarchical and k-means cluster analysis, the r-squared value, raw clustered data, processed clustered data of cluster centroids, and a ggplot object.

Details

Function to create a specified number of profiles of observed variables using a two-step (hierarchical and k-means) cluster analysis.

Examples

d <- pisaUSA15
m3 <- create_profiles_cluster(d,
                              broad_interest, enjoyment, instrumental_mot, self_efficacy,
                              n_profiles = 3)

#> Prepared data: Removed 354 incomplete cases

#> Hierarchical clustering carried out on: 5358 cases

#> K-means algorithm converged: 5 iterations

#> Clustered data: Using a 3 cluster solution

#> Calculated statistics: R-squared = 0.424

summary(m3)

#> # A tibble: 3 x 5
#>                 Cluster broad_interest enjoyment instrumental_mot self_efficacy
#>                   <chr>          <dbl>     <dbl>            <dbl>         <dbl>
#> 1 Profile 1 (2458 obs.)       2.829468  2.841131         2.322010      2.219372
#> 2 Profile 2 (1598 obs.)       3.177023  3.407541         1.395599      1.691776
#> 3 Profile 3 (1302 obs.)       1.683218  1.903418         2.430748      2.516520